sgroup-chatbot/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-genai>=1.0.0
mcp>=1.6.0
langgraph>=0.2.0
langchain-google-genai>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
feedparser>=6.0.11
